Support JS Handling from CefBaseApp
It is now possible to install JS a-sync routing directly into
CefBaseApp. A JS Handler is provided which can be installed as-is to
provide default dev_tools library functions.

If desired, it can be derived from and the virtual installer overriden
to install new functions. (One would expect in most cases the child
class would call back to the parent to pick up the dev_tools functions)