Force Time to be in UTC (as originally intended).
Previously Time initialised the gmt_offset to be 0, but this is not sufficient
to guarantee a UTC time after a call to mktime, which clobbers the input based
on the local timezone.

This has been fixed by using the timegm method. Whilst this is not portable,
dev_tools already assumes a Linux environment, so this is acceptable.