Pull in latest FixedJSONParserCPP Changes
Upgrades to the JSON library have been pulled back into dev_tools, since
it can't be submoudled until the cmake issue is resolved.

The upgrade delivers:
  - Improved test coverage (invalidJSON tests)
  - Support for the latest rapidjson version
  - Generator: Support ignoring NULL fields
  - Generator: Support merging of fields across objects in repeating lists
  - General Code Quality / Refactoring improvements

NOTE: It was necessary to retrofit timestamp support back in to the
FixedJSONParserCPP version, and to add support for trailing commas.

The port was validated by running the old test suite against the new
code. Finally the googletest suite from FixedJSONCPP was ported over,
with the Timestamp added back in.